Algorithm fast I guess
I have a hard time making the algorithm faster, as it has a lot to do with the garbage collector and making a lot of recursive calls, but I think that it hs to do with the speed/memory tradeoff that you usually get. Right now it uses a lot less memory, but that comes at the cost of some more expensive calls.

Update, simple integers are faster and uses less memory than nullable integers